Revert "Move sweeper FirebaseAndroidApp to the service package (#8117)" (#8132)
This reverts commit 5442ab1d1b605d04b651004bbeb00b35896e059f.
