+++ google/gcp_sweeper_test.go
@@ -6,8 +6,6 @@ import (
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-
-	_ "github.com/hashicorp/terraform-provider-google/google/services/firebase"
 )
 
 func TestMain(m *testing.M) {
