+++ google/resource_firebase_android_app_sweeper_test.go
@@ -1,6 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
-package firebase
+package google
 
 import (
 	"context"
